setup-env: bblayers.conf generated from manifest
That was tedious. Have <file>.bblayers live in conf/ directory where the
layers from the manifest, since one repository can contain multiple
layers (e.g. meta-openembedded), describing which "sub-" layer to use
for that manifest.

Also add some "python2 vs python3" logic in setup-env so bitbake/bin has
a link to the right version. Bitbake 1.28 isn't so happy with python3.

LCONF_VERSION has to be the right value as well (depending on the branch
version). Since this part of setup-env was changed add that as well.

Change-Id: I9e3fb9d31b060519c3f532af6ef9663d790c1db5
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>